test: event check details in form and details

+++ app/src/androidTest/java/org/dhis2/usescases/event/entity/EventDetailsUIModel.kt
@@ -2,7 +2,7 @@ package org.dhis2.usescases.event.entity
 
 data class EventDetailsUIModel (
     val programStage: String,
-    val completedPercentage: Float,
+    val completedPercentage: Int,
     val eventDate: String,
     val orgUnit: String
 )
